zuul: drop IE6-7 tests
IE6 and IE7 tests stopped working at some point, presumably because
some dependency of zuul or browserify added a dangling comma or some
feature that these browsers don't support.

I don't have the time/patience to debug this now. buffer probably works
in IE6-7 still, but I'm dropping the tests for it. If someone wants to
fix the issue and re-enable these tests, you're welcome to :)