Change error message when unknown encoding is given
The error message changed from

    "encoding" must be a valid string encoding

to

    Unknown encoding: <the-given-encoding>

To make the test work, parts of Node.js's `common` module need to
be used. Add the parts that are needed with minor modifications (the
check for the return code were removed, as vanilla JS errors don't
have an error code, only Node.js errors have).